+++ app/src/main/java/com/andrerinas/headunitrevived/aap/AapTransport.kt
@@ -195,6 +195,7 @@ class AapTransport(
         pollThread?.quit()
         sendThread?.quit()
         aapAudio.releaseAllFocus()
+        aapVideo.release()
 
         videoDecoder.onDecoderError = null